requirements.txt
flask
flask-socketio
python-socketio
google-genai
python-dotenv